refactor(runtime): import identity helpers directly
