+++ ddtrace/contrib/internal/subprocess/patch.py
@@ -15,8 +15,8 @@ from ddtrace.contrib import trace_utils
 from ddtrace.contrib.internal.subprocess.constants import COMMANDS
 from ddtrace.ext import SpanTypes
 from ddtrace.internal import core
+from ddtrace.internal._identity import get_runtime_propagation_envs
 from ddtrace.internal.logger import get_logger
-from ddtrace.internal.runtime import get_runtime_propagation_envs
 from ddtrace.internal.settings import env
 from ddtrace.internal.settings._config import config
 from ddtrace.internal.settings._telemetry import config as telemetry_config
